aarch64: testsuite: Update pr123748.c to use "long long" instead of "long"
Andrea pointed out that "long" is 32bits on mingw and the test case should use
"long long" instead.

Committed as obvious after testing.

gcc/testsuite/ChangeLog:

	* gcc.target/aarch64/pr123748.c: Use "long long" instead of "long".

+++ gcc/testsuite/gcc.target/aarch64/pr123748.c
@@ -2,7 +2,7 @@
 /* { dg-options "-Ofast" } */
 
 void
-f (float *__restrict f, long *__restrict l)
+f (float *__restrict f, long long *__restrict l)
 {
   for (int i = 0; i < 128; i++)
     f[i] = l[i];
